+++ tests/client-tests.js
@@ -7,7 +7,7 @@ import Hull from "../src";
 
 describe("Hull", () => {
   describe("as", () => {
-    it("should return scoped client with traits and track methods", () => {
+    it("should return scoped client with traits, track and alias methods", () => {
       const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
 
       const scopedAccount = hull.asAccount({ domain: "hull.io" });
@@ -17,10 +17,13 @@ describe("Hull", () => {
         .that.is.an("function");
       expect(scopedAccount).to.has.property("track")
         .that.is.an("function");
+      expect(scopedAccount).not.to.have.property("alias");
 
       expect(scopedUser).to.has.property("traits")
         .that.is.an("function");
       expect(scopedUser).to.has.property("track")
+        .that.is.an("function");
+      expect(scopedUser).to.has.property("alias")
         .that.is.an("function");
     });
 
@@ -36,6 +39,17 @@ describe("Hull", () => {
       expect(scopedJwtClaims)
         .to.have.property("io.hull.asUser")
         .that.eql({ email: "foo@bar.com" });
+    });
+
+    it("should expose an `as` method being an alias to `asUser`", () => {
+      const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
+
+      const scoped = hull.as("123456");
+      const scopedConfig = scoped.configuration();
+      const scopedJwtClaims = jwt.decode(scopedConfig.accessToken, scopedConfig.secret);
+      expect(scopedJwtClaims)
+        .to.have.property("sub")
+        .that.eql("123456");
     });
 
     it("should allow to pass user id as a string", () => {
@@ -125,6 +139,15 @@ describe("Hull", () => {
         .to.not.throw(Error);
       expect(hull.asAccount.bind(null, { external_id: "1234" }))
         .to.not.throw(Error);
+    });
+
+     it("should filter all non standard claims", () => {
+      const hull = new Hull({ id: "562123b470df84b740000042", secret: "1234", organization: "test" });
+
+      const scoped = hull.asUser({ email: "foo@bar.com", foo: "bar" });
+      const scopedJwtClaims = jwt.decode(scoped.configuration().accessToken, scoped.configuration().secret);
+      expect(scopedJwtClaims["io.hull.asUser"])
+        .to.eql({ email: "foo@bar.com" });
     });
   });
 });